+++ bel/src/environment.rs
@@ -39,41 +39,17 @@ impl Environment {
                 if list.is_empty() {
                     return Ok(Object::Symbol("nil".to_string()));
                 }
+                // if this list starts with a symbol, it may be 'special'
                 if let Object::Symbol(name) = list[0].clone() {
                     match name.as_ref() {
                         "set" => {
-                            for i in 1..list.len() - 1 {
-                                if let Object::Symbol(key) = list[i].clone() {
-                                    self.global.insert(key, list[i + 1].clone());
-                                } else {
-                                    return Err(BelError::InvalidObject {
-                                        expected: "symbol".to_string(),
-                                        found: list[i].t(),
-                                    });
-                                }
-                            }
-                            // append nil if the final arg isn't present
-                            // an even number of entries (including 'set')
-                            // means the last value is unspecified
-                            if list.len() % 2 == 0 {
-                                let i = list.len() - 1;
-                                if let Object::Symbol(key) = list[i].clone() {
-                                    self.global.insert(key, Object::Symbol("nil".to_string()));
-                                } else {
-                                    return Err(BelError::InvalidObject {
-                                        expected: "symbol".to_string(),
-                                        found: list[i].t(),
-                                    });
-                                }
-                            }
-                            return Ok(Object::Symbol("nil".to_string()));
+                            return self.set(list);
                         }
-                        // return the inner object without evaluating
-                        "quote" => {
-                            if list.len() != 2 {
-                                return Err(BelError::InvalidQuote(format!("{:?}", list)));
-                            }
-                            return Ok(list[1].clone());
+                       "quote" => {
+                            return self.quote(list);
+                        }
+                        "id" => {
+                            return self.id(&list[1..]);
                         }
                         _ => {}
                     }
@@ -93,6 +69,65 @@ impl Environment {
 
         Ok(output)
     }
+
+    fn set(&mut self, list: &Vec<Object>) -> Result<Object, BelError> {
+        for i in 1..list.len() - 1 {
+            if let Object::Symbol(key) = list[i].clone() {
+                self.global.insert(key, list[i + 1].clone());
+            } else {
+                return Err(BelError::InvalidObject {
+                    expected: "symbol".to_string(),
+                    found: list[i].t(),
+                });
+            }
+        }
+        // append nil if the final arg isn't present
+        // an even number of entries (including 'set')
+        // means the last value is unspecified
+        if list.len() % 2 == 0 {
+            let i = list.len() - 1;
+            if let Object::Symbol(key) = list[i].clone() {
+                self.global.insert(key, Object::Symbol("nil".to_string()));
+            } else {
+                return Err(BelError::InvalidObject {
+                    expected: "symbol".to_string(),
+                    found: list[i].t(),
+                });
+            }
+        }
+        Ok(Object::Symbol("nil".to_string()))
+    }
+
+    fn quote(&self, list: &[Object]) -> Result<Object, BelError> {
+       if list.len() == 2 {
+            // return the inner object without evaluating
+            Ok(list[1].clone())
+        } else {
+            Err(BelError::InvalidQuote(format!("{:?}", list)))
+        }        
+    }
+
+    fn id(&mut self, params: &[Object]) -> Result<Object, BelError> {
+        // id is true if 
+        // * there are two arguments
+        // * they are both symbols
+        // they have the same name
+        let mut result = Object::Symbol("nil".to_string());
+        if params.len() == 2 {
+            let arg0 = self.evaluate(&params[0])?;
+            if let Object::Symbol(lhs) = arg0 {
+                let arg1 = self.evaluate(&params[1])?;
+                if let Object::Symbol(rhs) = arg1 {
+                    if lhs == rhs {
+                        result = Object::Symbol("t".to_string());
+                    }
+                }
+            }
+        }
+            
+        Ok(result)
+    }
+    
 }
 
 #[cfg(test)]
@@ -157,4 +192,21 @@ mod tests {
 
         Ok(())
     }
+
+    #[test]
+    fn id_checks_for_identity() -> Result<(), BelError> {
+        let mut parser = parser::Parser::new();
+        let mut env = Environment::new();
+
+        let parse_obj = parser.parse("(id 'a 'a)")?;
+        let obj = env.evaluate(&parse_obj)?;
+        assert!(obj.is_true());
+
+        let parse_obj = parser.parse("(id 'a 'b)")?;
+        let obj = env.evaluate(&parse_obj)?;
+        assert!(obj.is_nil());
+
+        Ok(())
+    }
+
 }
